Move fake creation to extension.
