+++ src/xcode/ENA/ENA/Source/Services/TicketValidation/Models/TicketValidationInitializationData.swift
@@ -13,26 +13,4 @@ struct TicketValidationInitializationData: Codable {
 	let consent: String
 	let subject: String
 	let serviceProvider: String
-	
-	static func fake (
-		`protocol`: String = "protocol",
-		protocolVersion: String = "protocolVersion",
-		serviceIdentity: String = "serviceIdentity",
-		privacyUrl: String = "privacyUrl",
-		token: String = "token",
-		consent: String = "consent",
-		subject: String = "subject",
-		serviceProvider: String = "serviceProvider"
-	) -> TicketValidationInitializationData {
-		return TicketValidationInitializationData(
-			protocol: `protocol`,
-			protocolVersion: protocolVersion,
-			serviceIdentity: serviceIdentity,
-			privacyUrl: privacyUrl,
-			token: token,
-			consent: consent,
-			subject: subject,
-			serviceProvider: serviceProvider
-		)
-	}
 }
